Fixed notifications queue + compilation warning
- when showWithStatus was called while another notification was being
shown, it was wrongly changing its icon to spinner and when the first
action was finished, the new action would not see its status set
correctly. showWithStatus and showImage now share the same behaviour:
they add the action to the queue (or modify its status if already
exists) then if the action is primary show it, otherwise return. That
action will then be later picked up if it is still running.
- the animationDuration variable in positionBar was used without being
initialised. Set it to some default value.